Using the Custom Publisher
