+++ nats-test/src/events/base-publisher.ts
@@ -1,0 +1,22 @@
+import { Stan } from 'node-nats-streaming';
+import { Subjects } from './subjects';
+
+interface Event {
+    subject: Subjects;
+    data: any;
+}
+
+export abstract class Publisher<T extends Event> {
+    abstract subject: T['subject'];
+    private client: Stan;
+
+    constructor(client: Stan) {
+        this.client = client;
+    }
+
+    publish(data: T['data']) {
+        this.client.publish(this.subject, JSON.stringify(data), () => {
+            console.log('Event published.');
+        });
+    }
+}
