+++ nats-test/src/events/ticket-created-listener.ts
@@ -1,17 +1,19 @@
-import { Listener } from './base-listeners';
 import { Message } from 'node-nats-streaming';
+import { Listener } from './base-listener';
 import { TicketCreatedEvent } from './ticket-created-event';
 import { Subjects } from './subjects';
 
-
 export class TicketCreatedListener extends Listener<TicketCreatedEvent> {
-    readonly subject: Subjects.TicketCreated = Subjects.TicketCreated;
-    queueGroupName = 'payment-service';
+  readonly subject = Subjects.TicketCreated;
+  queueGroupName = 'payments-service';
 
-    onMessage(data: TicketCreatedEvent['data'], msg: Message): void {
-        console.log('Event data!', data);
+  onMessage(data: TicketCreatedEvent['data'], msg: Message) {
+    console.log('Event data!', data);
 
-        msg.ack();
-    }
+    console.log(data.id);
+    console.log(data.title);
+    console.log(data.price);
 
+    msg.ack();
+  }
 }
