+++ nats-test/src/events/ticket-created-publisher.ts
@@ -1,0 +1,7 @@
+import { Publisher } from './base-publisher';
+import { TicketCreatedEvent } from './ticket-created-event';
+import { Subjects } from './subjects';
+
+export class TicketCreatedPublisher extends Publisher<TicketCreatedEvent> {
+  subject: Subjects.TicketCreated = Subjects.TicketCreated;
+}
