+++ nats-test/src/listener.ts
@@ -5,24 +5,22 @@ import { TicketCreatedListener } from './events/ticket-created-listener';
 console.clear();
 
 const stan = nats.connect('ticketing', randomBytes(4).toString('hex'), {
-    url: 'http://localhost:4222',
+  url: 'http://localhost:4222',
 });
 
-
 stan.on('connect', () => {
-    console.log('Listener connected to NATS');
+  console.log('Listener connected to NATS');
 
-    stan.on('close', () => {
-        console.log('NATS connection closed!');
-        process.exit();
-    })
+  stan.on('close', () => {
+    console.log('NATS connection closed!');
+    process.exit();
+  });
 
-    new TicketCreatedListener(stan).listen();
+  new TicketCreatedListener(stan).listen();
 });
 
 process.on('SIGINT', () => stan.close());
 process.on('SIGTERM', () => stan.close());
-
