+++ nats-test/src/publisher.ts
@@ -1,4 +1,5 @@
 import nats from 'node-nats-streaming';
+import { TicketCreatedPublisher } from './events/ticket-created-publisher';
 
 console.clear();
 
@@ -9,13 +10,20 @@ const stan = nats.connect('ticketing', 'abc', {
 stan.on('connect', () => {
   console.log('Publisher connected to NATS');
 
-  const data = JSON.stringify({
+  const publisher = new TicketCreatedPublisher(stan);
+  publisher.publish({
     id: '123',
     title: 'concert',
     price: 20,
   });
 
-  stan.publish('ticket:created', data, () => {
-    console.log('Event published');
-  });
+  // const data = JSON.stringify({
+  //   id: '123',
+  //   title: 'concert',
+  //   price: '$20',
+  // });
+
+  // stan.publish('TicketCreated', data, () => {
+  //   console.log('Event published');
+  // });
 });
